requirements.txt
loguru
selenium
tenacity
tqdm
openpyxl
pyyaml
python-dotenv
